테스트 기능 완성 exam 저장 완성

+++ exams/models.py
@@ -6,7 +6,7 @@ class Exam(models.Model):
     exam_type = models.CharField(max_length=10, default='en')  # 문제타입 - en = 한글뜻 보이고 영어 맞추기, ko = 영어 보이고 한글 맞추기
     in_word = models.CharField(max_length=50)
     exam_right = models.BooleanField()
-    exam_point = models.DecimalField(max_digits=3, decimal_places=2)
+    exam_point = models.FloatField()
     exam_difficulty = models.IntegerField(default=1)
     reg_date = models.DateTimeField(auto_now_add=True)
     # auth.User 는 Django 에 자동 생성 되는 User 테이블
